backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
anthropic==0.40.0
python-multipart==0.0.9
Pillow==10.4.0
python-dotenv==1.0.1
pytz==2024.1
